% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{EconomicIndicatorsGUI}
\alias{EconomicIndicatorsGUI}
\title{EconomicIndicatorsGUI}
\usage{
EconomicIndicatorsGUI()
}
\value{
shiny app
}
\description{
Runs a Graphical User Interface to plot GDP, CPI and Unemployment Rate time series of 12 European Countries.
}
\details{
Data span from 1983 to 2018.
Selected countries are: Austria, Belgium, Denmark, France, Germany, Ireland, Italy, Netherlands, Portugal, Spain, Sweden, UK.
}
