% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{GDP}
\format{dataframe}
\source{
\url{https://fred.stlouisfed.org/}
}
\usage{
GDP
}
\description{
A dataset containing 12 European countries' GDP time ordered observations.
}
\details{
Data span from 1983 to 2018.
The countries are: Austria, Belgium, Denmark, France, Germany, Ireland, Italy, Netherlands, Portugal, Spain, Sweden, United Kingdom.
}
\keyword{datasets}
